# packages/contracts/openapi/admin-api.yaml
openapi: 3.1.0
info:
  title: Admin API
  version: 0.4.4
  description: |
    API behind the admin application (Store view + HQ view). Every request carries a Keycloak staff-realm JWT.
    Store-scoped resources live under `/admin/stores/{storeId}/...`; the server checks the OpenFGA relation named
    in each operation's `x-permission` against `store:{storeId}` (or `organization:hq`) before doing anything.
    UI gating is a convenience; these checks are the security boundary. Every mutation writes `audit_log`.
    Money is `{ amount_minor, currency }`. Ids are uuids. Frozen at tag contracts-v0.1; 0.2.0 adds optional sort/order to list operations (additive);
    0.3.0 adds the marketing surface under `/admin/stores/{storeId}/marketing/…` and `/admin/marketing/…` (docs/marketing-scope.md, additive).
    0.4.0 adds search merchandising rules under `/admin/stores/{storeId}/merchandising/…` (CONTRACT CHANGE #162, additive) and
    documents `401` and `403` on every operation (#180): every `x-permission` can fail, and Prism can only mock what is written down.
    0.4.1 adds per-item product media + signed Cloudinary upload params (#168), `getPromotion`/`updatePromotion` with the
    `buy_x_get_y` type and `stackable`/`exclusive` (#189), and spells `ProductFeed` out so `status: error` validates (#194).
servers:
  - url: http://localhost:4011
    description: Prism mock (pnpm mock)
  - url: https://admin-api.example.com/v1
security:
  - staffToken: []
...
paths:
...
  /admin/stores/{storeId}/orders/{orderId}/refunds:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
      - { $ref: '#/components/parameters/OrderId' }
    post:
      tags: [orders]
      operationId: createRefund
      summary: Refund part or all of a captured payment. `support` may refund up to the store's support_refund_limit_minor.
      x-permission: { relation: support, object: 'store:{storeId}' }
      parameters:
        - {
            name: Idempotency-Key,
            in: header,
            required: true,
            schema: { type: string, minLength: 8 },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_minor, reason]
              properties:
                payment_id:
                  { type: string, format: uuid, description: Defaults to the captured payment }
                amount_minor: { type: integer, minimum: 1 }
                reason: { type: string, enum: [return, cancellation, goodwill, chargeback] }
                return_id: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Refund' }
              example:
                {
                  id: 60000000-0000-4000-8000-000000000001,
                  order_id: 30000000-0000-4000-8000-000000000501,
                  payment_id: 60000000-0000-4000-8000-000000000101,
                  return_id: null,
                  amount: { amount_minor: 500, currency: EUR },
                  reason: goodwill,
                  status: succeeded,
                  provider_refund_id: re_mock,
                  created_at: '2026-09-04T11:00:00Z',
                }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    staffToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak staff realm token
  parameters:
    StoreId:
      name: storeId
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: unauthorized, message: Missing bearer token, details: {} }
    Forbidden:
      description: Permission check failed (OpenFGA)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            {
              code: forbidden,
              message: 'requires finance on organization:hq',
              details: { relation: finance, object: 'organization:hq' },
            }
...
    Conflict:
      description: Uniqueness or state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: conflict, message: handle already exists, details: { field: handle } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          {
            type: string,
            description: 'validation_error, unauthorized, forbidden, not_found, conflict, out_of_stock, payment_failed, internal',
          }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Money:
      type: object
      required: [amount_minor, currency]
      properties:
        amount_minor: { type: integer }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
...
    Refund:
      type: object
      required:
        [
          id,
          order_id,
          payment_id,
          return_id,
          amount,
          reason,
          status,
          provider_refund_id,
          created_at,
        ]
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        payment_id: { type: string, format: uuid }
        return_id: { type: [string, 'null'], format: uuid }
        amount: { $ref: '#/components/schemas/Money' }
        reason: { type: string, enum: [return, cancellation, goodwill, chargeback] }
        status: { type: string, enum: [pending, succeeded, failed] }
        provider_refund_id: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
...